Restructure update (#96)
* Update README.adoc

* Update README.adoc

* Revert "Update README.adoc"

This reverts commit 8c76a5c01e72d8a2b62d2c07330158dc553927a9.

* Revert "Update README.adoc"

This reverts commit e004ce954f78120947aeafe6e5610654aa2de00f.

* Update README.adoc

* Updating hotspots

* Revert "Updating hotspots"

This reverts commit 815fda87c2dfffdebfc880e53d03ad13c9eab649.

* Update PersonServiceIT.3.java

* Update README.adoc

* Update README.adoc

* Remove redundant PersonServiceIT

* Update README.adoc

* Update README.adoc

* Add hotspots for removed code

* Revert "Add hotspots for removed code"

This reverts commit ec2c5390520c03387a6cfd6637b694c437af779f.

* Update README.adoc

* Correct linter issues

* Fix linter issues

* Correct linter issues

* Revert "Correct linter issues"

This reverts commit 91285afcae3b4638f4cbd156981d4893fcb92567.

* Update copyright

* Add hotspots for removal

* Update README.adoc

* Update README.adoc

* Update README.adoc

* Update copyright

* Fix linter issues

* Adding subsections

* Revert section

* Update README.adoc

* Update README.adoc

* Update README.adoc

* Update README.adoc

* Revert "Update README.adoc"

This reverts commit 39aa5aa7dfa2ecf2921aeef0bf31f5186b172d12.

* Revert "Update README.adoc"

This reverts commit 1759d573ad6d023bbb6a42815b2b310352424a22.

* Revert "Update README.adoc"

This reverts commit dd35730b7f392ec36c8b7299002e96ca151bc013.

* Revert "Update README.adoc"

This reverts commit 81c48700038aa27f0a25216b4980890e69797b60.

* Revert "Revert section"

This reverts commit bea4d071e40857c0239e770e28a851cbe160ad81.

* Fix hotspots

* Revert first test section

* Update README.adoc

* Address incorrect code on cloud

* Update README.adoc

* update versions
