+++ finish/src/main/liberty/config/server.xml
@@ -4,7 +4,7 @@
         <feature>jaxrs-2.1</feature>
         <feature>jsonb-1.0</feature>
         <!-- tag::mpHealth[] -->
-        <feature>mpHealth-3.0</feature>
+        <feature>mpHealth-3.1</feature>
         <!-- end::mpHealth[] -->
         <feature>mpConfig-2.0</feature>
         <feature>mpRestClient-2.0</feature>
