+++ hotspots/src/test/java/io/openliberty/guides/testing/ErrorPathIT.java
@@ -1,22 +1,14 @@
 // tag::copyright[]
-/*
- * Copyright (c) 2019 IBM Corporation and others
+/*******************************************************************************
+ * Copyright (c) 2019, 2021 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
  *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * You may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+ * Contributors:
+ *     IBM Corporation - Initial implementation
+ *******************************************************************************/
 // end::copyright[]
 package io.openliberty.guides.testing;
 
@@ -68,14 +60,14 @@ public class ErrorPathIT {
 
     @Test
     public void testCreateBadPersonNegativeAge() {
-        assertThrows(BadRequestException.class, () -> 
+        assertThrows(BadRequestException.class, () ->
           personSvc.createPerson("NegativeAgePersoN", -1));
     }
 
     @Test
     public void testCreateBadPersonNameTooLong() {
-        assertThrows(BadRequestException.class, () -> 
-          personSvc.createPerson("NameTooLongPersonNameTooLongPersonNameTooLongPerson", 
+        assertThrows(BadRequestException.class, () ->
+          personSvc.createPerson("NameTooLongPersonNameTooLongPersonNameTooLongPerson",
           5));
     }
 }
