+++ start/src/main/liberty/config/server.xml
@@ -3,7 +3,7 @@
     <featureManager>
         <feature>jaxrs-2.1</feature>
         <feature>jsonb-1.0</feature>
-        <feature>mpHealth-3.0</feature>
+        <feature>mpHealth-3.1</feature>
         <feature>mpConfig-2.0</feature>
         <feature>mpRestClient-2.0</feature>
         <feature>beanValidation-2.0</feature>
